refactor: reconstruct folder and add barrel file
