+++ src/features/floating-write-btn/floating-write-btn.tsx
@@ -1,27 +1,21 @@
 'use client'
 
-import { useAuth } from '@/src/shared/provider'
 import { Pencil, X } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-
-interface FloatingButtonProps {
-  onClick: () => void
-  text: string
-}
+import { FloatingMenuButton } from '@/src/features'
 
 interface FloatingWriteButtonProps {
   isClick?: boolean
   setIsClick?: (value: boolean) => void
 }
 
-export default function FloatingWriteButton({
+export function FloatingWriteButton({
   isClick: isClickProp,
   setIsClick: setIsClickProp,
 }: FloatingWriteButtonProps) {
   const router = useRouter()
   const [internalIsClick, setInternalIsClick] = useState(false)
-  const { token } = useAuth()
 
   const isClick = isClickProp ?? internalIsClick
   const setIsClick = setIsClickProp ?? setInternalIsClick
@@ -32,10 +26,6 @@ export default function FloatingWriteButton({
     }
     document.scrollingElement?.scrollTo({ top: 0, behavior: 'smooth' })
     router.push(path)
-  }
-
-  if (!token) {
-    return null
   }
 
   return (
@@ -68,16 +58,5 @@ export default function FloatingWriteButton({
         </button>
       </div>
     </div>
-  )
-}
-
-export function FloatingMenuButton({ onClick, text }: FloatingButtonProps) {
-  return (
-    <button
-      className='w-[80px] h-[30px] bg-white shadow-grid text-[14px] rounded-full flex items-center justify-center hover:bg-brand hover:text-white transition-all duration-200'
-      onClick={onClick}
-    >
-      {text}
-    </button>
   )
 }
