+++ src/features/index.ts
@@ -12,3 +12,4 @@ export * from './header-elements'
 export * from './firebase'
 export * from './comment-card'
 export * from './user'
+export * from './floating-write-btn'
