+++ src/views/list-course/index.tsx
@@ -6,9 +6,8 @@ import { useGetCourses } from '@/src/entities/course'
 import useRegionStore, { LikeRegion } from '@/src/shared/store/regionStore'
 import CourseListLayout from '@/src/widgets/course-list-layout'
 import { Spacer, SelectCategories, useToast } from '@/src/shared/ui'
-import FloatingWriteButton from '@/src/widgets/floating-write-btn'
 import { useDeleteMyLikeRegion, usePostMyLikeRegion } from '@/src/entities/user'
-import { SelectSort } from '@/src/features'
+import { SelectSort, FloatingWriteButton } from '@/src/features'
 import { useAuth } from '@/src/shared/provider'
 
 export default function ListCourse() {
