+++ src/views/list-plan/index.tsx
@@ -1,13 +1,16 @@
 'use client'
 
 import { useEffect, useState } from 'react'
-import FloatingWriteButton from '@/src/widgets/floating-write-btn'
 import { Spacer, ProfileImage } from '@/src/shared/ui'
 import { PlanType, useGetPlans, NoPlanToolTip } from '@/src/entities/plan'
 import { getLoginUrl } from '@/src/entities/auth'
 import { useRouter } from 'next/navigation'
 import { useGetMyProfile } from '@/src/entities/user'
-import { CoursePlanCard, SkeletonCoursePlanCard } from '@/src/features'
+import {
+  CoursePlanCard,
+  SkeletonCoursePlanCard,
+  FloatingWriteButton,
+} from '@/src/features'
 import { SkeletonDiv } from '@/src/shared/ui'
 
 export default function ListPlan() {
